+++ angular/test/test-app/e2e/src/modal.spec.ts
@@ -60,4 +60,8 @@ describe('Modals: Inline', () => {
     cy.get('ion-list ion-item:nth-child(3)').should('have.text', 'C');
     cy.get('ion-list ion-item:nth-child(4)').should('have.text', 'D');
   });
+
+  it('should have a div with .ion-page', () => {
+    cy.get('ion-modal').children('.ion-page').should('exist');
+  });
 });