+++ packages/react/test-app/src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
-import { IonApp, setupIonicReact } from '@ionic/react';
+import { IonApp, IonButton, IonContent, IonHeader, IonPage, IonRouterOutlet, IonTitle, IonToolbar, setupIonicReact } from '@ionic/react';
 import { IonReactRouter } from '@ionic/react-router';
 
 /* Core CSS required for Ionic components to work properly */
@@ -30,9 +30,11 @@ setupIonicReact();
 const App: React.FC = () => (
   <IonApp>
     <IonReactRouter>
-      <Route path="/" component={Main} />
-      <Route path="/overlay-hooks" component={OverlayHooks} />
-      <Route path="/overlay-components" component={OverlayComponents} />
+      <IonRouterOutlet>
+        <Route path="/" component={Main} />
+        <Route path="/overlay-hooks" component={OverlayHooks} />
+        <Route path="/overlay-components" component={OverlayComponents} />
+      </IonRouterOutlet>
     </IonReactRouter>
   </IonApp>
 );