+++ packages/vue/test-app/src/App.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script lang="ts">
-import { IonApp, IonRouterOutlet } from '@ionic/vue';
+import { IonApp, IonRouterOutlet, useIonRouter } from '@ionic/vue';
 import { defineComponent } from 'vue';
 
 export default defineComponent({
@@ -13,6 +13,9 @@ export default defineComponent({
   components: {
     IonApp,
     IonRouterOutlet
+  },
+  setup() {
+    (window as any).debugIonRouter = useIonRouter();
   }
 });
 </script>